+++ packages/stream_chat/lib/src/core/models/user.g.dart
@@ -25,6 +25,7 @@ User _$UserFromJson(Map<String, dynamic> json) {
     teams:
         (json['teams'] as List<dynamic>?)?.map((e) => e as String).toList() ??
             [],
+    language: json['language'] as String? ?? 'en',
   );
 }
 
@@ -47,5 +48,6 @@ Map<String, dynamic> _$UserToJson(User instance) {
   writeNotNull('online', readonly(instance.online));
   writeNotNull('banned', readonly(instance.banned));
   val['extra_data'] = instance.extraData;
+  writeNotNull('language', readonly(instance.language));
   return val;
 }