+++ packages/stream_chat/CHANGELOG.md
@@ -1,3 +1,13 @@
+## Upcoming
+
+🛑️ Removed
+- The `MessageTranslation` class has been removed. Use the new `i18n` field in the `Message` class instead.
+
+✅ Added
+
+- The `Message` class now has an `i18n` field for translations
+- The `User` class now has a `language` field for the user's language preference.
+
 ## 2.0.0
 
 🛑️ Breaking Changes from `1.5.3`